requirements.txt
python-docx
transformers
